% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{batch_get_resource_config}
\alias{batch_get_resource_config}
\title{Returns the current configuration for one or more requested resources}
\usage{
batch_get_resource_config(resourceKeys)
}
\arguments{
\item{resourceKeys}{A list of resource keys to be processed with the current request. Each element in the list consists of the resource type and resource ID.}
}
\description{
Returns the current configuration for one or more requested resources. The operation also returns a list of resources that are not processed in the current request. If there are no unprocessed resources, the operation returns an empty unprocessedResourceKeys list.
}
\details{
\itemize{
\item The API does not return results for deleted resources.
\item The API does not return any tags for the requested resources. This information is filtered out of the supplementaryConfiguration section of the API response.
}
}
\examples{

}
