% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{delete_configuration_aggregator}
\alias{delete_configuration_aggregator}
\title{Deletes the specified configuration aggregator and the aggregated data associated with the aggregator}
\usage{
delete_configuration_aggregator(ConfigurationAggregatorName)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}
}
\description{
Deletes the specified configuration aggregator and the aggregated data associated with the aggregator.
}
\section{Accepted Parameters}{
\preformatted{delete_configuration_aggregator(
  ConfigurationAggregatorName = "string"
)
}
}

