% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{delete_evaluation_results}
\alias{delete_evaluation_results}
\title{Deletes the evaluation results for the specified AWS Config rule}
\usage{
delete_evaluation_results(
  ConfigRuleName = "string"
)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the AWS Config rule for which you want to delete the evaluation results.}
}
\description{
Deletes the evaluation results for the specified AWS Config rule. You can specify one AWS Config rule per request. After you delete the evaluation results, you can call the StartConfigRulesEvaluation API to start evaluating your AWS resources against the rule.
}
\examples{

}
