% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{delete_pending_aggregation_request}
\alias{delete_pending_aggregation_request}
\title{Deletes pending authorization requests for a specified aggregator account in a specified region}
\usage{
delete_pending_aggregation_request(RequesterAccountId, RequesterAwsRegion)
}
\arguments{
\item{RequesterAccountId}{The 12-digit account ID of the account requesting to aggregate data.}

\item{RequesterAwsRegion}{The region requesting to aggregate data.}
}
\description{
Deletes pending authorization requests for a specified aggregator account in a specified region.
}
\examples{

}
