% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{deliver_config_snapshot}
\alias{deliver_config_snapshot}
\title{Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel}
\usage{
deliver_config_snapshot(deliveryChannelName)
}
\arguments{
\item{deliveryChannelName}{[required] The name of the delivery channel through which the snapshot is delivered.}
}
\description{
Schedules delivery of a configuration snapshot to the Amazon S3 bucket in the specified delivery channel. After the delivery has started, AWS Config sends the following notifications using an Amazon SNS topic that you have specified.
}
\details{
\itemize{
\item Notification of the start of the delivery.
\item Notification of the completion of the delivery, if the delivery was successfully completed.
\item Notification of delivery failure, if the delivery failed.
}
}
\section{Accepted Parameters}{
\preformatted{deliver_config_snapshot(
  deliveryChannelName = "string"
)
}
}

