% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_aggregate_compliance_by_config_rules}
\alias{describe_aggregate_compliance_by_config_rules}
\title{Returns a list of compliant and noncompliant rules with the number of resources for compliant and noncompliant rules}
\usage{
describe_aggregate_compliance_by_config_rules(ConfigurationAggregatorName,
  Filters = NULL, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{Filters}{Filters the results by ConfigRuleComplianceFilters object.}

\item{Limit}{The maximum number of evaluation results returned on each page. The default is maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns a list of compliant and noncompliant rules with the number of resources for compliant and noncompliant rules.
}
\details{
The results can return an empty result page, but if you have a nextToken, the results are displayed on the next page.
}
\section{Accepted Parameters}{
\preformatted{describe_aggregate_compliance_by_config_rules(
  ConfigurationAggregatorName = "string",
  Filters = list(
    ConfigRuleName = "string",
    ComplianceType = "COMPLIANT"|"NON_COMPLIANT"|"NOT_APPLICABLE"|"INSUFFICIENT_DATA",
    AccountId = "string",
    AwsRegion = "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

