% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_config_rule_evaluation_status}
\alias{describe_config_rule_evaluation_status}
\title{Returns status information for each of your AWS managed Config rules}
\usage{
describe_config_rule_evaluation_status(ConfigRuleNames = NULL,
  NextToken = NULL, Limit = NULL)
}
\arguments{
\item{ConfigRuleNames}{The name of the AWS managed Config rules for which you want status information. If you do not specify any names, AWS Config returns status information for all AWS managed Config rules that you use.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}

\item{Limit}{The number of rule evaluation results that you want returned.

This parameter is required if the rule limit for your account is more than the default of 50 rules.

For information about requesting a rule limit increase, see \href{http://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_config}{AWS Config Limits} in the \emph{AWS General Reference Guide}.}
}
\description{
Returns status information for each of your AWS managed Config rules. The status includes information such as the last time AWS Config invoked the rule, the last time AWS Config failed to invoke the rule, and the related error for the last failure.
}
\examples{

}
