% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_configuration_recorder_status}
\alias{describe_configuration_recorder_status}
\title{Returns the current status of the specified configuration recorder}
\usage{
describe_configuration_recorder_status(
  ConfigurationRecorderNames = list(
    "string"
  )
)
}
\arguments{
\item{ConfigurationRecorderNames}{The name(s) of the configuration recorder. If the name is not specified, the action returns the current status of all the configuration recorders associated with the account.}
}
\description{
Returns the current status of the specified configuration recorder. If a configuration recorder is not specified, this action returns the status of all configuration recorders associated with the account.
}
\details{
Currently, you can specify only one configuration recorder per region in your account.
}
\examples{

}
