% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_configuration_recorders}
\alias{describe_configuration_recorders}
\title{Returns the details for the specified configuration recorders}
\usage{
describe_configuration_recorders(
  ConfigurationRecorderNames = list(
    "string"
  )
)
}
\arguments{
\item{ConfigurationRecorderNames}{A list of configuration recorder names.}
}
\description{
Returns the details for the specified configuration recorders. If the configuration recorder is not specified, this action returns the details for all configuration recorders associated with the account.
}
\details{
Currently, you can specify only one configuration recorder per region in your account.
}
\examples{

}
