% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_delivery_channels}
\alias{describe_delivery_channels}
\title{Returns details about the specified delivery channel}
\usage{
describe_delivery_channels(DeliveryChannelNames = NULL)
}
\arguments{
\item{DeliveryChannelNames}{A list of delivery channel names.}
}
\description{
Returns details about the specified delivery channel. If a delivery channel is not specified, this action returns the details of all delivery channels associated with the account.
}
\details{
Currently, you can specify only one delivery channel per region in your account.
}
\section{Accepted Parameters}{
\preformatted{describe_delivery_channels(
  DeliveryChannelNames = list(
    "string"
  )
)
}
}

