% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{describe_pending_aggregation_requests}
\alias{describe_pending_aggregation_requests}
\title{Returns a list of all pending aggregation requests}
\usage{
describe_pending_aggregation_requests(Limit = NULL, NextToken = NULL)
}
\arguments{
\item{Limit}{The maximum number of evaluation results returned on each page. The default is maximum. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns a list of all pending aggregation requests.
}
\examples{

}
