% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{get_aggregate_compliance_details_by_config_rule}
\alias{get_aggregate_compliance_details_by_config_rule}
\title{Returns the evaluation results for the specified AWS Config rule for a specific resource in a rule}
\usage{
get_aggregate_compliance_details_by_config_rule(ConfigurationAggregatorName,
  ConfigRuleName, AccountId, AwsRegion, ComplianceType = NULL, Limit = NULL,
  NextToken = NULL)
}
\arguments{
\item{ConfigurationAggregatorName}{The name of the configuration aggregator.}

\item{ConfigRuleName}{The name of the AWS Config rule for which you want compliance information.}

\item{AccountId}{The 12-digit account ID of the source account.}

\item{AwsRegion}{The source region from where the data is aggregated.}

\item{ComplianceType}{The resource compliance status.

For the \code{GetAggregateComplianceDetailsByConfigRuleRequest} data type, AWS Config supports only the \code{COMPLIANT} and \code{NON_COMPLIANT}. AWS Config does not support the \code{NOT_APPLICABLE} and \code{INSUFFICIENT_DATA} values.}

\item{Limit}{The maximum number of evaluation results returned on each page. The default is 50. You cannot specify a number greater than 100. If you specify 0, AWS Config uses the default.}

\item{NextToken}{The nextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns the evaluation results for the specified AWS Config rule for a specific resource in a rule. The results indicate which AWS resources were evaluated by the rule, when each resource was last evaluated, and whether each resource complies with the rule.
}
\details{
The results can return an empty result page. But if you have a nextToken, the results are displayed on the next page.
}
\examples{

}
