% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{get_compliance_details_by_resource}
\alias{get_compliance_details_by_resource}
\title{Returns the evaluation results for the specified AWS resource}
\usage{
get_compliance_details_by_resource(ResourceType, ResourceId,
  ComplianceTypes = NULL, NextToken = NULL)
}
\arguments{
\item{ResourceType}{The type of the AWS resource for which you want compliance information.}

\item{ResourceId}{The ID of the AWS resource for which you want compliance information.}

\item{ComplianceTypes}{Filters the results by compliance.

The allowed values are \code{COMPLIANT}, \code{NON_COMPLIANT}, and \code{NOT_APPLICABLE}.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns the evaluation results for the specified AWS resource. The results indicate which AWS Config rules were used to evaluate the resource, when each rule was last used, and whether the resource complies with each rule.
}
\examples{

}
