% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{get_compliance_summary_by_resource_type}
\alias{get_compliance_summary_by_resource_type}
\title{Returns the number of resources that are compliant and the number that are noncompliant}
\usage{
get_compliance_summary_by_resource_type(ResourceTypes = NULL)
}
\arguments{
\item{ResourceTypes}{Specify one or more resource types to get the number of resources that are compliant and the number that are noncompliant for each resource type.

For this request, you can specify an AWS resource type such as \code{AWS::EC2::Instance}. You can specify that the resource type is an AWS account by specifying \code{AWS::::Account}.}
}
\description{
Returns the number of resources that are compliant and the number that are noncompliant. You can specify one or more resource types to get these numbers for each resource type. The maximum number returned is 100.
}
\examples{

}
