% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{get_resource_config_history}
\alias{get_resource_config_history}
\title{Returns a list of configuration items for the specified resource}
\usage{
get_resource_config_history(resourceType, resourceId, laterTime = NULL,
  earlierTime = NULL, chronologicalOrder = NULL, limit = NULL,
  nextToken = NULL)
}
\arguments{
\item{resourceType}{The resource type.}

\item{resourceId}{The ID of the resource (for example., \code{sg-xxxxxx}).}

\item{laterTime}{The time stamp that indicates a later time. If not specified, current time is taken.}

\item{earlierTime}{The time stamp that indicates an earlier time. If not specified, the action returns paginated results that contain configuration items that start when the first configuration item was recorded.}

\item{chronologicalOrder}{The chronological order for configuration items listed. By default, the results are listed in reverse chronological order.}

\item{limit}{The maximum number of configuration items returned on each page. The default is 10. You cannot specify a number greater than 100. If you specify 0, AWS Config uses the default.}

\item{nextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns a list of configuration items for the specified resource. The list contains details about each state of the resource during the specified time interval. If you specified a retention period to retain your \code{ConfigurationItems} between a minimum of 30 days and a maximum of 7 years (2557 days), AWS Config returns the \code{ConfigurationItems} for the specified retention period.
}
\details{
The response is paginated. By default, AWS Config returns a limit of 10 configuration items per page. You can customize this number with the \code{limit} parameter. The response includes a \code{nextToken} string. To get the next page of results, run the request again and specify the string for the \code{nextToken} parameter.

Each call to the API is limited to span a duration of seven days. It is likely that the number of records returned is smaller than the specified \code{limit}. In such cases, you can make another call, using the \code{nextToken}.
}
\examples{

}
