% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{list_discovered_resources}
\alias{list_discovered_resources}
\title{Accepts a resource type and returns a list of resource identifiers for the resources of that type}
\usage{
list_discovered_resources(resourceType, resourceIds = NULL,
  resourceName = NULL, limit = NULL, includeDeletedResources = NULL,
  nextToken = NULL)
}
\arguments{
\item{resourceType}{The type of resources that you want AWS Config to list in the response.}

\item{resourceIds}{The IDs of only those resources that you want AWS Config to list in the response. If you do not specify this parameter, AWS Config lists all resources of the specified type that it has discovered.}

\item{resourceName}{The custom name of only those resources that you want AWS Config to list in the response. If you do not specify this parameter, AWS Config lists all resources of the specified type that it has discovered.}

\item{limit}{The maximum number of resource identifiers returned on each page. The default is 100. You cannot specify a number greater than 100. If you specify 0, AWS Config uses the default.}

\item{includeDeletedResources}{Specifies whether AWS Config includes deleted resources in the results. By default, deleted resources are not included.}

\item{nextToken}{The \code{nextToken} string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Accepts a resource type and returns a list of resource identifiers for the resources of that type. A resource identifier includes the resource type, ID, and (if available) the custom resource name. The results consist of resources that AWS Config has discovered, including those that AWS Config is not currently recording. You can narrow the results to include only resources that have specific resource IDs or a resource name.
}
\details{
You can specify either resource IDs or a resource name, but not both, in the same request.

The response is paginated. By default, AWS Config lists 100 resource identifiers on each page. You can customize this number with the \code{limit} parameter. The response includes a \code{nextToken} string. To get the next page of results, run the request again and specify the string for the \code{nextToken} parameter.
}
\examples{

}
