% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{put_aggregation_authorization}
\alias{put_aggregation_authorization}
\title{Authorizes the aggregator account and region to collect data from the source account and region}
\usage{
put_aggregation_authorization(
  AuthorizedAccountId = "string",
  AuthorizedAwsRegion = "string"
)
}
\arguments{
\item{AuthorizedAccountId}{[required] The 12-digit account ID of the account authorized to aggregate data.}

\item{AuthorizedAwsRegion}{[required] The region authorized to collect aggregated data.}
}
\description{
Authorizes the aggregator account and region to collect data from the source account and region.
}
\examples{

}
