% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{put_configuration_aggregator}
\alias{put_configuration_aggregator}
\title{Creates and updates the configuration aggregator with the selected source accounts and regions}
\usage{
put_configuration_aggregator(
  ConfigurationAggregatorName = "string",
  AccountAggregationSources = list(
    list(
      AccountIds = list(
        "string"
      ),
      AllAwsRegions = TRUE|FALSE,
      AwsRegions = list(
        "string"
      )
    )
  ),
  OrganizationAggregationSource = list(
    RoleArn = "string",
    AwsRegions = list(
      "string"
    ),
    AllAwsRegions = TRUE|FALSE
  )
)
}
\arguments{
\item{ConfigurationAggregatorName}{[required] The name of the configuration aggregator.}

\item{AccountAggregationSources}{A list of AccountAggregationSource object.}

\item{OrganizationAggregationSource}{An OrganizationAggregationSource object.}
}
\description{
Creates and updates the configuration aggregator with the selected source accounts and regions. The source account can be individual account(s) or an organization.
}
\details{
AWS Config should be enabled in source accounts and regions you want to aggregate.

If your source type is an organization, you must be signed in to the master account and all features must be enabled in your organization. AWS Config calls \code{EnableAwsServiceAccess} API to enable integration between AWS Config and AWS Organizations.
}
\examples{

}
