% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{put_configuration_recorder}
\alias{put_configuration_recorder}
\title{Creates a new configuration recorder to record the selected resource configurations}
\usage{
put_configuration_recorder(ConfigurationRecorder)
}
\arguments{
\item{ConfigurationRecorder}{The configuration recorder object that records each configuration change made to the resources.}
}
\description{
Creates a new configuration recorder to record the selected resource configurations.
}
\details{
You can use this action to change the role \code{roleARN} or the \code{recordingGroup} of an existing recorder. To change the role, call the action on the existing configuration recorder and specify a role.

Currently, you can specify only one configuration recorder per region in your account.

If \code{ConfigurationRecorder} does not have the \strong{recordingGroup} parameter specified, the default is to record all supported resource types.
}
\examples{

}
