% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{start_configuration_recorder}
\alias{start_configuration_recorder}
\title{Starts recording configurations of the AWS resources you have selected to record in your AWS account}
\usage{
start_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{The name of the recorder object that records each configuration change made to the resources.}
}
\description{
Starts recording configurations of the AWS resources you have selected to record in your AWS account.
}
\details{
You must have created at least one delivery channel to successfully start the configuration recorder.
}
\examples{

}
