% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{stop_configuration_recorder}
\alias{stop_configuration_recorder}
\title{Stops recording configurations of the AWS resources you have selected to record in your AWS account}
\usage{
stop_configuration_recorder(ConfigurationRecorderName)
}
\arguments{
\item{ConfigurationRecorderName}{The name of the recorder object that records each configuration change made to the resources.}
}
\description{
Stops recording configurations of the AWS resources you have selected to record in your AWS account.
}
\examples{

}
