% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.configservice_operations.R
\name{delete_retention_configuration}
\alias{delete_retention_configuration}
\title{Deletes the retention configuration}
\usage{
delete_retention_configuration(RetentionConfigurationName)
}
\arguments{
\item{RetentionConfigurationName}{The name of the retention configuration to delete.}
}
\description{
Deletes the retention configuration.
}
\examples{

}
